% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_to_monthly.R
\name{aggregate_to_monthly}
\alias{aggregate_to_monthly}
\title{Aggregates the sub-monthly data in a riversCentralAsia data set to monthly
data.}
\usage{
aggregate_to_monthly(dataTable, funcTypeLib)
}
\arguments{
\item{dataTable}{A tibble of the format \code{ChirchikRiverBasin}. Must contain at
least the columns date, data, norm, type and code.}

\item{funcTypeLib}{is a list of functions with associated data types that
will be applied to the data. Currently, the aggregation functions
\code{mean} and \code{sum} are supported. The user specifies the
data types which are to be aggregated with either \code{mean} or
\code{sum}.}
}
\value{
Returns a tibble of the same format as \code{data} with data
            aggregated to monthly time steps and "mon" in the resolution
            column.
            Returns 1 if aggregation fails.
}
\description{
A riversCentralAsia data set can be of type discharge, precipitation, and
temperature whereby temerature may be mean, minimum or maximum. The function
aggregate_data aggregates all sub-monthly data in the data tibble to monthly
data, according to the type-function pair given in the second argument.
The user specifies which aggregation function to use for each data type using
timetk's summarise_by_time.
}
\examples{
dataTable <- ChirchikRiverBasin
funcTypeLib <- list(mean = c("Q", "T"), sum= "P")
data_mon <- aggregate_to_monthly(dataTable, funcTypeLib)

}
\seealso{
Other Helper functions: 
\code{\link{\%>\%}()},
\code{\link{assess_fc_qual}()},
\code{\link{cutRaster2Basin}()},
\code{\link{decadeMaker}()},
\code{\link{generateHourlyFromDaily_PT}()},
\code{\link{generateSeqDates}()},
\code{\link{hyear}()},
\code{\link{monDateSeq}()},
\code{\link{monthlyNorms_ClimateScenario}()},
\code{\link{posixct2rsminerveChar}()}
}
\concept{Helper functions}
