% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AizensGlacierFunctions.R
\name{AizensGlacierFunctions}
\alias{AizensGlacierFunctions}
\title{Estimates glacier surface area and glacier volume.}
\usage{
AizensGlacierFunctions(glaciers)
}
\arguments{
\item{glaciers}{A data.frame, tibble, or sf object. Must contain an attribute
```Lmax``` in meters (as in the Randolph Glacier Inventory (RGI 6.0)) or
```Length [km]``` in kilometers or ```L(t) [km]``` (as in the output from
Oerlemans glacier length function).}
}
\value{
Returns the input object with Area_Aizen_km2 and Volume_Aizen_km3
  added. Will return NULL if the length column in the input object is not
  found.
}
\description{
AizensGlacierFunctions implements the empirical relationships between glacier
length, surface area and volume found by Aizen et al. in the Tien Shan
mountains. They can be used to estimate regional glacier volumes.
}
\examples{
glacier <- tibble::tibble(Year = c(2000:2010),
                          "Length [km]" = c(5.7, 5.6, 5.5, 5.5, 5.4, 5.3,
                                            5.3, 5.2, 5.1, 5, 4.8))
AizensGlacierFunctions(glacier)
}
\references{
Aizen, Aizen & Kuzmichonok (2007) Glaciers and hydrological
  changes in the Tien Shan: simulation and prediction. Environmental Research
  Letters. DOI: 10.1088/1748-9326/2/4/045019.
}
\seealso{
Other Glacier functions: 
\code{\link{ErazovsGlacierFunction}()},
\code{\link{OerlemansGlacierLengthModel_FormatLong}()},
\code{\link{OerlemansGlacierLengthModel}()},
\code{\link{glacierArea_Erasov}()},
\code{\link{glacierArea_RGIF}()},
\code{\link{glacierBalanceP}()},
\code{\link{glacierBalance}()},
\code{\link{glacierDischargeRMSE}()},
\code{\link{glacierDischarge_HM}()},
\code{\link{glacierImbalAbl}()},
\code{\link{glacierMelt_TI}()},
\code{\link{glacierRMSE}()},
\code{\link{glacierTotalAblation_HM}()},
\code{\link{glacierVolume_Erasov}()},
\code{\link{glacierVolume_RGIF}()}
}
\concept{Glacier functions}
