% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChirchikRiverBasin}
\alias{ChirchikRiverBasin}
\title{Hydrometeorological Data from the Chirchik River Basin, Central Asia.}
\format{
A data frame with columns:
\describe{
 \item{date}{A date value.}
 \item{data}{The actual data as specified by 'type'.}
 \item{norm}{Norm of data as calculated over entire span of timeseries.}
 \item{units}{Data units corresponding to the 'type' of data.}
 \item{type}{Factor with levels Q, P, and T where Q is for discharge, P is for
 precipitation and T is for temperature.}
 \item{code}{Official 5-digit code of hydrological or meteorological station.}
 \item{station}{Name of station or village where station is located.}
 \item{river}{Name of tributary river.}
 \item{basin}{Name of river basin.}
 \item{resolution}{Factor describing temporal resolution of data, either
 10-days decadal 'dec' or monthly 'mon'.}
 \item{lon_UTM42}{Longitude of station in UTM 42 coordinates.}
 \item{lat_UTM42}{Latitude of station in UTM 42 coordinates.}
 \item{altitude_masl}{Station elevation in meters above sea level.}
 \item{basinSize_sqkm}{Size of upstream contributing catchment.}
}
}
\source{
Central Asia Hydrometeorological Agencies
}
\usage{
ChirchikRiverBasin
}
\description{
Hydrometeorological Data from the Chirchik River Basin, Central Asia.
}
\examples{
\dontrun{
 ChirchikRiverBasin
}
}
\keyword{datasets}
