% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climateScenarioPreparation_RMAWGEN.R
\name{climateScenarioPreparation_RMAWGEN}
\alias{climateScenarioPreparation_RMAWGEN}
\title{Climate Scenario Preparations}
\usage{
climateScenarioPreparation_RMAWGEN(pathN, fileN, param)
}
\arguments{
\item{pathN}{Path to climate scenario files that are downscaled on catchment-subbasin elevation bands}

\item{fileN}{File name where climate projections are stored (.rda file)}

\item{param}{RMAWGEN parameters stored as list}
}
\value{
List of 12 climate scenarios, i.e. RCP45 and RCP85, climate models ACCESS1_3, CMCC-CM and MIROC5, two periods of interest (2051-2060 and 2091-2100). Scenario name, monthly precipitation (prec), minimum (tasmin) and maximum (tasmax) temperatures, year_min and year_max are stored in scenario list entires.
}
\description{
Climate scenarios output from function `downscale_ClimPred_monthly_BasinElBands()` is prepared for the stochastic weather generator RMAWGEN.
}
