% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeAnnualFlowDurationCurve.R
\name{computeAnnualFlowDurationCurve}
\alias{computeAnnualFlowDurationCurve}
\title{Computes the annual flow duration curve of a river discharge data set.}
\usage{
computeAnnualFlowDurationCurve(data, column, date = Date)
}
\arguments{
\item{data}{Tibble with date and discharge in columns.}

\item{column}{Name of the column with data to calculate duration curve from.}

\item{date}{Name of the date column}
}
\value{
Input tibble with sorted data and duration stats in columns.
}
\description{
Computes the annual flow duration curve of a river discharge data set.
}
\examples{
Qdf <- tibble::tibble(
  Date = seq.Date(from = lubridate::as_date("2020-01-01"),
                  to = lubridate::as_date("2022-12-13"),
                  by = "month"),
  Q = rep(c(1:6, 6:1), 3)
)
DurationCurve <- computeAnnualFlowDurationCurve(Qdf, "Q", "Date")
}
