% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeDiurnalTemperatureCycle.R
\name{computeDiurnalTemperatureCycle}
\alias{computeDiurnalTemperatureCycle}
\title{Using ERA5 hourly data, the function computes the diurnal temperature cycle for each station.}
\usage{
computeDiurnalTemperatureCycle(era5_data, param)
}
\arguments{
\item{era5_data}{ERA5 csv-file (RS MINERVE Database input) with details on stations and hourly P,T and Q data.}

\item{param}{Parameter list with year_min and year_max fields that denote the period of interest for which the mean diurnal per-sttion cycles should be computed.}
}
\value{
Tibble with average diurnal temperature cycles stored in each column for each station.
}
\description{
The function takes the tabular ERA5 input data and computes the diurnal cycle for a period of interest between param$year_min and param$yearmax
}
