% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutRaster2Basin.R
\name{cutRaster2Basin}
\alias{cutRaster2Basin}
\title{Cutting raster file to basin shape with reprojection}
\usage{
cutRaster2Basin(rasterIn, aoiRegion_latlon, aoiBasin_UTM, proj_UTM)
}
\arguments{
\item{rasterIn}{Raster file (e.g. DEM)}

\item{aoiRegion_latlon}{Extent object of the larger region (i.e. for Central Asia, a good choice is extent(c(65,80.05,35.95,44.05))).}

\item{aoiBasin_UTM}{Extent object of the basin shapefile}

\item{proj_UTM}{CRS of UTM projection (i.e. for 42N in Central Asia, use "+init=epsg:32642")}
}
\value{
Projected raster cut to basin
}
\description{
This function cuts a raster dataset to a specific basin shapefiles and project the raster
according to the specified projection.
}
\seealso{
Other Helper functions: 
\code{\link{\%>\%}()},
\code{\link{aggregate_to_monthly}()},
\code{\link{decadeMaker}()},
\code{\link{generateSeqDates}()},
\code{\link{hyear}()},
\code{\link{monDateSeq}()},
\code{\link{posixct2rsminerveChar}()}
}
\concept{Helper functions}
