% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacierArea_Erasov.R
\name{glacierArea_Erasov}
\alias{glacierArea_Erasov}
\title{Calculates glacier area from glacier volume}
\usage{
glacierArea_Erasov(volume_km3)
}
\arguments{
\item{volume_km3}{Glacier volume in km3}
}
\value{
Glacier area in km2
}
\description{
Inverse of Erasov glacier scaling function. Only works for positive glacier
volume.
}
\seealso{
\code{\link{glacierVolume_Erasov}}, \code{\link{glacierArea_RGIF}}

Other glacier functions: 
\code{\link{glacierArea_RGIF}()},
\code{\link{glacierBalance}()},
\code{\link{glacierDischarge_HM}()},
\code{\link{glacierImbalAbl}()},
\code{\link{glacierMelt_TI}()},
\code{\link{glacierVolume_Erasov}()},
\code{\link{glacierVolume_RGIF}()}
}
\concept{glacier functions}
