% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacierArea_RGIF.R
\name{glacierArea_RGIF}
\alias{glacierArea_RGIF}
\title{Calculates glacier area based on glacier volume}
\source{
RGI v6.0 <https://www.glims.org/RGI/>,
  Farinotti et al., 2019 <https://doi.org/10.1038/s41561-019-0300-3>
}
\usage{
glacierArea_RGIF(volume_km3)
}
\arguments{
\item{volume_km3}{glacier volume in km3}
}
\value{
glacier area in km2
}
\description{
Scaling relationships derived from the RGI v6.0 region 13 data set, combined
with the glacier thickness derived by Farinotti et al. 2019.
}
\details{
Assuming a normal distribution of the residuals, the relative
  uncertainty of the volume estimate is given as 2 times the standard
  deviation of the relative residuals which is equal to 53%. This method of
  error estimation likely underestimates the actual uncertainty.
}
\examples{
  V_km3 <- c(0.1, 1, 10)
  A_km2 <- glacierArea_RGIF(V_km3)
}
\seealso{
\code{[glacierVolume_RGIF]}, \code{[glacierArea_Erasov]}

Other glacier functions: 
\code{\link{glacierArea_Erasov}()},
\code{\link{glacierBalanceP}()},
\code{\link{glacierBalance}()},
\code{\link{glacierDischargeRMSE}()},
\code{\link{glacierDischarge_HM}()},
\code{\link{glacierImbalAbl}()},
\code{\link{glacierMelt_TI}()},
\code{\link{glacierRMSE}()},
\code{\link{glacierTotalAblation_HM}()},
\code{\link{glacierVolume_Erasov}()},
\code{\link{glacierVolume_RGIF}()}
}
\concept{glacier functions}
