% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacierBalance.R
\name{glacierBalance}
\alias{glacierBalance}
\title{Calculate glacier mass balance of all glaciers in a basin over time.}
\usage{
glacierBalance(melt_a_eb, rgi_elbands)
}
\arguments{
\item{melt_a_eb}{Annual glacier melt in mm/a for each glacier/elevation band.
The aggregation period is from October 1 of the previous year to September
30 of the current year. The object \code{melt_a_eb} must contain the
columns \code{Hyear} for the hydrological year, \code{ID} for the
glacier ID in the RGI format and \code{Melt} with the glacier melt in mm/a.
The column \code{ID} is generated by paste0(RGIId, "_", elvtn_b) in the
input rgi_elbands.}

\item{rgi_elbands}{A shape file based on the Randolph Glacier Inventory but
with the larger glaciers subdivided into elevation bands. The object
\code{rgi_elbands} must contain the columns \code{RGIId}, \code{A_km2} with
the area of the glaciers/elevation bands in km2, \code{thickness_m} with
the thickness of the glaciers/elevation bands in m. We further need a
column called \code{Area_tot_glacier_km2} with the area of the entire
glacier within the catchment and a column \code{ID} which is consistent with
the \code{ID} column in object \code{melt_a_eb}.}
}
\value{
A tibble containing annual glacier melt,  imbalance ablation,
  glacier volume, glacier area.
}
\description{
Small glaciers are treated as one unit whereas large glaciers are treated in
elevation bands.
}
\details{
The input object \code{melt_a_eb} can be generated by aggregating the
  daily glacier melt obtained with the function \code{glacierMelt_TI()} to
  the hydrological year. The vignette \code{glaciers-02} demonstrates a use
  case for the function. \cr
  The function performs a simplified glacier mass balance on an annual time
  step \code{t} as follows:  \cr
    - Update the glacier area \code{A(t) = glacierArea_RGIF(V(t-1))} \cr
    - Calculate glacier discharge from glacier melt \code{Qgl(t) = M(t)*A(t)} \cr
    - Estimate the imbalance ablation from the glacier melt
      \code{QglImb(t) = glacierImbalAbl(M(t))} \cr
    - Update the glacier volume \code{V(t) = V(t-1) + QglImb(t)}
}
\seealso{
Other Glacier functions: 
\code{\link{AizensGlacierFunctions}()},
\code{\link{ErazovsGlacierFunction}()},
\code{\link{OerlemansGlacierLengthModel_FormatLong}()},
\code{\link{OerlemansGlacierLengthModel}()},
\code{\link{glacierArea_Erasov}()},
\code{\link{glacierArea_RGIF}()},
\code{\link{glacierBalanceP}()},
\code{\link{glacierDischargeRMSE}()},
\code{\link{glacierDischarge_HM}()},
\code{\link{glacierImbalAbl}()},
\code{\link{glacierMelt_TI}()},
\code{\link{glacierRMSE}()},
\code{\link{glacierTotalAblation_HM}()},
\code{\link{glacierVolume_Erasov}()},
\code{\link{glacierVolume_RGIF}()}
}
\concept{Glacier functions}
