% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacierDischarge_HM.R
\name{glacierDischarge_HM}
\alias{glacierDischarge_HM}
\title{Estimates the total glacier discharge based on glacier thinning rates}
\usage{
glacierDischarge_HM(dhdt)
}
\arguments{
\item{dhdt}{Glacier thinning rates in meters per year (m/a).
dhdt < 0: glacier is loosing mass,
dhdt > 0: glacier is gaining mass.}
}
\value{
Total glacier melt in m3/a.
  Negative for glacier melt and positive for glacier gain.
}
\description{
Regression derived from the data set by Hugonnet et al., 2021 and Miles et
al., 2021.
}
\seealso{
Other glacier functions: 
\code{\link{glacierArea_Erasov}()},
\code{\link{glacierArea_RGIF}()},
\code{\link{glacierImbalAbl}()},
\code{\link{glacierMelt_TI}()},
\code{\link{glacierVolume_Erasov}()},
\code{\link{glacierVolume_RGIF}()}
}
\concept{glacier functions}
