% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacierDischargeRMSE.R
\name{glacierDischargeRMSE}
\alias{glacierDischargeRMSE}
\title{Calculates glacier melt and the root mean squared error between observed
glacier melt and simulated glacier melt.}
\usage{
glacierDischargeRMSE(
  parameters,
  temperature,
  observed,
  index = 1:(length(unique(observed$RGIID)))
)
}
\arguments{
\item{parameters}{The two parameters of the function ```glacierMelt_TI```.
The first parameter is the melt factor ```MF```, the second parameter is
the threshold temperature ```threshold_temperature``` above which melt sets
on. Example: c(4, 0).}

\item{temperature}{A tibble with the time steps in rows and temperature
values for individual hydrological response units in columns. The names of
the hydrological response units is typically of the form RGIId_layer,
e.g. RGI60-13.10007_1. The RGIId and the underbar are required as the
algorithm splits the name at the underbar and mergest the observed glacier
melt by RGIId. \cr
The dates column of the temperature time series can be in the format
lubridate::date for sub-annual glacier melt simulations or an integer
indicating the (hydrological) year for which glacier melt is calculated.
The date column must be named ```date``` for sub-annual time steps or
```year``` for annual time steps.}

\item{observed}{A tibble with the glacier melt data by Miles et al.
```observed``` must contain the columns RGIID, totAbl and Area_m2. If a
column ```year``` is present, the algorithm assumes yearly data, else it
assumes that ```observed``` represents average data over the entire period
of the temperature time series.}

\item{index}{(optional) number to indicate for which individual HRU the RMSE
should be returned. Defaults to all.}
}
\value{
-1*RMSE between observed glacier melt and glacier melt simulated with
  the temperature index model of indexed hydrological response units.
}
\description{
This function calls ```glacierMelt_TI``` with ```parameters```
  and ```temperature``` and calculates the root mean squared error RMSE
  between the simulated and observed ```observed``` glacier melt. The
  optional ```index``` serves to limit the results to one single glacier. \
  This function can be used to calibrate the parameters of the temperature
  index model. As the recommended optimization algorithm GA is a maximiser,
  the function returns the negative RMSE.
}
\seealso{
\code{\link{glacierMelt_TI}}

Other Glacier functions: 
\code{\link{AizensGlacierFunctions}()},
\code{\link{ErazovsGlacierFunction}()},
\code{\link{OerlemansGlacierLengthModel_FormatLong}()},
\code{\link{OerlemansGlacierLengthModel}()},
\code{\link{glacierArea_Erasov}()},
\code{\link{glacierArea_RGIF}()},
\code{\link{glacierBalanceP}()},
\code{\link{glacierBalance}()},
\code{\link{glacierDischarge_HM}()},
\code{\link{glacierImbalAbl}()},
\code{\link{glacierMelt_TI}()},
\code{\link{glacierRMSE}()},
\code{\link{glacierTotalAblation_HM}()},
\code{\link{glacierVolume_Erasov}()},
\code{\link{glacierVolume_RGIF}()}
}
\concept{Glacier functions}
