% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacierImbalAbl.R
\name{glacierImbalAbl}
\alias{glacierImbalAbl}
\title{Estimates the imbalance ablation based on the total glacier discharge}
\usage{
glacierImbalAbl(totAbl_m3a)
}
\arguments{
\item{totAbl}{total glacier ablation (eq. glacier melt) in m3/a. A positive
number}
}
\value{
imbalAbl Imbalance ablation component of total ablation in m3/a.
  Negative for glacier melt and positive for glacier gain.
}
\description{
Regression derived from the data set by Miles et al., 2021. In our modelling
framework the glacier discharge is a positive number while in Miles glacier
balance model it is a negative number. Note that imbalance ablation computed
here is consistent with Miles glacier balance model: negative imbalance
ablation indicates glacier loss and positive glacier imbalance ablation
indicates glacier gain.
}
\seealso{
Other glacier functions: 
\code{\link{glacierArea_Erasov}()},
\code{\link{glacierArea_RGIF}()},
\code{\link{glacierVolume_Erasov}()},
\code{\link{glacierVolume_RGIF}()}
}
\concept{glacier functions}
