% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacierMelt_TI.R
\name{glacierMelt_TI}
\alias{glacierMelt_TI}
\title{Glacier melt computed with a temperature index model}
\source{
Hock R. (2003): Temperature index melt modelling in mountain areas.
  Journal of Hydrology 282, pp 104--115, DOI: 10.1016/S0022-1694(03)00257-9.
}
\usage{
glacierMelt_TI(temperature, MF = 4, threshold_temperature = 0)
}
\arguments{
\item{temperature}{A matrix with daily temperatures over time (rows) for each
glacier and/or elevation band (columns).}

\item{MF}{is a scalar or a named vector with temperature index factors which
contains one melt factor for each glacier/elevation band in the
temperature matrix.}

\item{threshold_temperature}{is a scalar or a named vector with the
temperature above which glacier melt is produced.}
}
\value{
A matrix with melt rates in mm per glacier/elevation band.
}
\description{
Glacier melt (\code{M}) is produced by multiplying the melt factor
  (\code{MF}) with the amount of energy available for melting which is
  approximated with the difference between the measured temperature (\code{T})
  and a threshold temperature above which melt occurs (\code{Tth}). \cr
  The formula is given as: \cr
  \code{M = MF * (T - Tth) if T is above Tth}. \cr
  No melt occurs at temperatures below the threshold temperature.
}
\examples{
# Generate random temperature forcing
number_of_glaciers <- 10
number_of_days <- 50*365
temperature <- matrix(
  runif(number_of_glaciers * number_of_days, min = -10, max = 10),
  nrow = number_of_days, ncol = number_of_glaciers)
colnames(temperature) <- paste0("Gl", 1:number_of_glaciers)
# Generate sample melt factors
MF <- temperature[1, ] * 0 + 1:number_of_glaciers
# Calculate glacier melt assuming a threshold temperature for glacier melt of
# 1 degree Celcius.
melt <- glacierMelt_TI(temperature, MF, threshold_temperature = 1)
}
\seealso{
Other glacier functions: 
\code{\link{glacierArea_Erasov}()},
\code{\link{glacierArea_RGIF}()},
\code{\link{glacierBalance}()},
\code{\link{glacierDischarge_HM}()},
\code{\link{glacierImbalAbl}()},
\code{\link{glacierRMSE}()},
\code{\link{glacierVolume_Erasov}()},
\code{\link{glacierVolume_RGIF}()}
}
\concept{glacier functions}
