% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacierTotalAblation_HM.R
\name{glacierTotalAblation_HM}
\alias{glacierTotalAblation_HM}
\title{Estimates the total glacier ablation based on glacier mass loss rates.}
\usage{
glacierTotalAblation_HM(dmdtda)
}
\arguments{
\item{dmdtda}{Glacier mass loss rates in meters of water equivalents per year
(m/a).  \cr
dhdt < 0: glacier is loosing mass, \cr
dhdt > 0: glacier is gaining mass. \cr}
}
\value{
Total glacier ablation in m3/a.
  Negative for glacier melt and positive for glacier gain.
}
\description{
Regression derived from the data set by Hugonnet et al., 2021 (dmdtda) and
Miles et al., 2021 (totAbl).
}
\seealso{
Other glacier functions: 
\code{\link{glacierArea_Erasov}()},
\code{\link{glacierArea_RGIF}()},
\code{\link{glacierBalance}()},
\code{\link{glacierDischarge_HM}()},
\code{\link{glacierImbalAbl}()},
\code{\link{glacierMelt_TI}()},
\code{\link{glacierRMSE}()},
\code{\link{glacierVolume_Erasov}()},
\code{\link{glacierVolume_RGIF}()}
}
\concept{glacier functions}
