% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacierVolume_Erasov.R
\name{glacierVolume_Erasov}
\alias{glacierVolume_Erasov}
\title{Calculates glacier volume from glacier area.}
\usage{
glacierVolume_Erasov(area_km2)
}
\arguments{
\item{area_km2}{Glacier area in km2}
}
\value{
Glacier volume in km3
}
\description{
Empirical scaling function to calculate glacier volume based on glacier area
by Erasov, 1968. Only works for positive glacier area.
}
\seealso{
\code{\link{glacierArea_Erasov}}, \code{\link{glacierVolume_RGIF}}

Other Glacier functions: 
\code{\link{ErazovsGlacierFunction}()},
\code{\link{glacierAreaVolume_Aizen}()},
\code{\link{glacierArea_Erasov}()},
\code{\link{glacierArea_RGIF}()},
\code{\link{glacierMelt_TI}()},
\code{\link{glacierVolume_RGIF}()}
}
\concept{Glacier functions}
