% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacierVolume_RGIF.R
\name{glacierVolume_RGIF}
\alias{glacierVolume_RGIF}
\title{Calculates glacier volume based on glacier area}
\usage{
glacierVolume_RGIF(area_km2)
}
\arguments{
\item{area_km2}{glacier area in km2}
}
\value{
glacier volume in km3
}
\description{
Scaling relationships derived from the RGI v6.0 region 13 data set, combined
with the glacier thickness derived by Farinotti et al. 2019.
}
\seealso{
\code{\link{glacierArea_RGIF}}, \code{\link{glacierVolume_Erasov}}

Other glacier functions: 
\code{\link{glacierArea_Erasov}()},
\code{\link{glacierArea_RGIF}()},
\code{\link{glacierBalance}()},
\code{\link{glacierDischarge_HM}()},
\code{\link{glacierImbalAbl}()},
\code{\link{glacierMelt_TI}()},
\code{\link{glacierVolume_Erasov}()}
}
\concept{glacier functions}
