% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_iEasyHydro_data.R
\name{import_iEasyHydro_data}
\alias{import_iEasyHydro_data}
\title{Import iEasyHydro Data}
\usage{
import_iEasyHydro_data(
  fPath,
  fName,
  code,
  stationName,
  rName,
  rBasin,
  dataType,
  units,
  x_UTM42N,
  y_UTM42N,
  masl,
  basin_size
)
}
\arguments{
\item{fPath}{Path to the input file}

\item{fName}{File name}

\item{code}{Hydrometeorological station code}

\item{stationName}{Hydrometeorological station name/location}

\item{rName}{Name of river}

\item{rBasin}{Name of basin}

\item{dataType}{Type of data, either `Q` (discharge data), `T` (temperature data) or `P` (precipitation data)}

\item{units}{Data units}

\item{x_UTM42N}{x coordinates in UTM42N}

\item{y_UTM42N}{y coordinates in UTM42N}

\item{masl}{station elevation}

\item{basin_size}{basin size in km^2}
}
\value{
Time-aware tibble with relevant data
}
\description{
Loads csv files that were exported from iEasyHydro. The function automatically computes
the long-term norm of the data provided and returns a time aware tibble with
date, data, data norm columns and code columns, etc. This function is an adapted version
of the riversCentralAsis::loadTabularData() function.
}
