% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_minerve_input_csv.R
\name{load_minerve_input_csv}
\alias{load_minerve_input_csv}
\title{Reads climate and discharge data into a tibble.}
\usage{
load_minerve_input_csv(filename)
}
\arguments{
\item{filename}{Path to file with input data for RS Minerve.}
}
\value{
Returns a tibble of the same format as the inbuilt \code{data} in
  ChirchikRiverBasin with columns \code{date} (POSIXct), \code{basin} (char),
  \code{type} (char, T, P or Q), \code{unit} (char, C, mm/a or m3/s) and
  \code{value} (num). Data is in hourly (climate) to decadal (10 days) or
  monthly (discharge) time steps.
}
\description{
The data is in a csv file that can also be imported to RS Minerve to form a
database.
}
\examples{
\dontrun{
load_minerve_input_csv("path_to_file.csv")
}

}
\author{
Beatrice Marti, hydrosolutions
}
