% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTabularData.R
\name{loadTabularData}
\alias{loadTabularData}
\title{Load Tabular Hydro-Meteorological Data}
\usage{
loadTabularData(fPath, fName, code)
}
\arguments{
\item{fPath}{Path to the input file}

\item{fName}{File name}

\item{code}{Station code}
}
\value{
A matrix of the infile
}
\description{
Loads csv files with tabular hydrometerological data where years are in rows
and decades or months are in columns. The function automatically detects if
monthly or decadal (10-day) data is provided. The function automatiaclly computes
the long-term norm of the data provided and returns a time aware tibble with
date, data, data norm columns and code columns.
}
