% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monDateSeq.R
\name{monDateSeq}
\alias{monDateSeq}
\title{Create a monthly sequence date vector with end of month day dates}
\usage{
monDateSeq(st, en, freq)
}
\arguments{
\item{st}{start date}

\item{en}{end date}

\item{freq}{frequence of dates (12 for continuous monthly dates)}
}
\value{
A sequence of dates
}
\description{
This function creates a sequence of monthly dates between a start and an end
date. It requires the specification of the start and end dates and the frequency.
If a continuous sequence is requested, then freq should be 12.
}
\seealso{
Other Helper functions: 
\code{\link{\%>\%}()},
\code{\link{aggregate_to_monthly}()},
\code{\link{cutRaster2Basin}()},
\code{\link{decadeMaker}()},
\code{\link{generateHourlyFromDaily_PT}()},
\code{\link{generateSeqDates}()},
\code{\link{hyear}()},
\code{\link{posixct2rsminerveChar}()}
}
\concept{Helper functions}
