% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monthlyNorms_ClimateScenario.R
\name{monthlyNorms_ClimateScenario}
\alias{monthlyNorms_ClimateScenario}
\title{Compute Monthly Norm of Min. and Max. Temperature and Precipitation of a Climate Scenario}
\usage{
monthlyNorms_ClimateScenario(climScenario)
}
\arguments{
\item{climScenario}{Climate scenario (see also function climateScenarioPreparation_RMAWGEN())}
}
\value{
climScenario list with added norms (prec_norm, tasmin_norm, tasmax_norm)
}
\description{
Helper function to compute monthly norms of climate variables
}
