% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipe.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe graphics}
\description{
Like dplyr, tidET also uses the pipe function, \code{\%>\%} to turn
function composition into a series of imperative statements.
}
\seealso{
Other Helper functions: 
\code{\link{aggregate_to_monthly}()},
\code{\link{cutRaster2Basin}()},
\code{\link{decadeMaker}()},
\code{\link{generateSeqDates}()},
\code{\link{hyear}()},
\code{\link{monDateSeq}()},
\code{\link{posixct2rsminerveChar}()}
}
\concept{Helper functions}
