% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotAnnualFlowDurationCurve.R
\name{plotAnnualFlowDurationCurve}
\alias{plotAnnualFlowDurationCurve}
\title{plot the annual flow duration curve of a river discharge data set.}
\usage{
plotAnnualFlowDurationCurve(data, column)
}
\arguments{
\item{data}{tibble output of computeAnnualFlowDurationCurve}

\item{column}{str name of column with discharge data to plot (same as in
computeAnnualFlowDurationCurve()).}
}
\value{
A figure with the yearly average rating curve (black line) and the 
  data of the individual years in grey.
}
\description{
plot the annual flow duration curve of a river discharge data set.
}
\seealso{
[computeAnnualFlowDurationCurve()]
}
