% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotNormDevHYY.R
\name{plotNormDevHYY}
\alias{plotNormDevHYY}
\title{Helper function to plot value deviations from norms for entire hydrological years and
cold and warm seasons.}
\usage{
plotNormDevHYY(data2Plot, dataType, stationID)
}
\arguments{
\item{data2Plot}{tibble}

\item{dataType}{String, either 'Q' for mean discharge in m3/s, 'T' for temperature in degrees Celsius or 'P' for precipitation in mm/yr.}

\item{stationID}{String of station name, e.g. 'Khorog-Gunt 17050'}
}
\value{
ggplot object
}
\description{
The function expects a tibble columns 'hyYear' (date), data (dbl), data_cs (dbl) and data_ws (dbl)
where hyYear is a date column and data, data_cs and data_ws columns are the values for the entire year,
the cold season (Oct. to Mar. in Central Asia) and warm season (Apr. to Sept. in Central Asia). The following
dataType(s) are accepted: 'Q' for discharge, 'T' temperature in deg. C. and 'P' for precipitation in mm/year.
The data2Plot tibble can easily be produced with the function convert2HYY() that is part of the package.
}
\seealso{
Other Pre-processing: 
\code{\link{biasCorrect_ERA5_CHELSA}()},
\code{\link{climateScenarioPreparation_RMAWGEN}()},
\code{\link{computeDiurnalTemperatureCycle}()},
\code{\link{doQuantileMapping}()},
\code{\link{downscale_ClimPred_monthly_BasinElBands}()},
\code{\link{gen_HRU_Climate_CSV_RSMinerve}()},
\code{\link{gen_basinElevationBands}()},
\code{\link{generate_ERA5_Subbasin_CSV}()},
\code{\link{import_iEasyHydro_data}()},
\code{\link{loadTabularData}()},
\code{\link{prepare_RMAWGEN_input_data}()}
}
\concept{Pre-processing}
