% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readForcingSTR.R
\name{readForcingSTR}
\alias{readForcingSTR}
\title{Reads climate and discharge data from a character string into a tibble.}
\usage{
readForcingSTR(string, tz = "UTC")
}
\arguments{
\item{string}{Character string with input data for RS Minerve.}

\item{tz}{Time zone character to be passed to lubridate::as_datetime.
Defaults to "UTC"}
}
\value{
Returns a tibble of the same format as \code{data} with data
        in hourly (climate) to decadal or monthly (discharge) time steps.
        Includes all attributes of the csv file.
}
\description{
The data is in a string in the same format as the RSMinerve forcing csv
that can also be imported to RS Minerve to form a database. This function is
useful for reformatting data in character format that is to be written to
RS Minerve input files.
}
\seealso{
Other RS Minerve IO: 
\code{\link{getChunkSize}()},
\code{\link{load_minerve_input_csv}()},
\code{\link{readDBCSV}()},
\code{\link{readForcingCSV}()},
\code{\link{readRSMParameters}()},
\code{\link{readResultCSV}()},
\code{\link{readResultDST}()},
\code{\link{readSelectionCHK}()},
\code{\link{translateCSVtoDST}()},
\code{\link{writeRSMParameters}()},
\code{\link{writeSelectionCHK}()}
}
\concept{RS Minerve IO}
