% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readRSMParameters.R
\name{readRSMParameters}
\alias{readRSMParameters}
\title{Reads parameters from RSMinerve parameter file}
\usage{
readRSMParameters(filepath)
}
\arguments{
\item{filepath}{Character with path to file to read.}
}
\value{
Tibble containing all parameters read from file.
}
\description{
Reads parameters from RSMinerve parameter file
}
\details{
The suported RSMInerve objects are described in the vignette
  Overview Objects. \cr
  `parameters` contains 6 columns: \cr
    `Object` (character) identifying the RSMinerve object type (e.g.
      Station). See Vignette Parameters for more information on the objects
      available in RSMinerveR. \cr
    `Name` is the user specified name of the object in the RSMinerve model
      (e.g. Meteo station A). \cr
    `Zone` is the name of the zone an object is assigned to (e.g. "A"). \cr
    `Parameters` names all parameters for each object (e.g. X \[m\], Y \[m\],
      etc.). The parameter names are the same as in the RS Minerve parameter
      file. See Vignette Parameters for more information on the parameters of
      the available objects in RSMinerveR. \cr
    `Values` contains the parameter values as numerics (e.g. 4500, 3000).
}
\examples{
\dontrun{
filepath <- normalizePath(file.path("Tutorial_Parameters.txt"))
params <- readRSMParameters(filepath)
}
}
\seealso{
Other RS Minerve IO: 
\code{\link{getChunkSize}()},
\code{\link{load_minerve_input_csv}()},
\code{\link{readDBCSV}()},
\code{\link{readForcingCSV}()},
\code{\link{readForcingSTR}()},
\code{\link{readResultCSV}()},
\code{\link{readResultDST}()},
\code{\link{readSelectionCHK}()},
\code{\link{translateCSVtoDST}()},
\code{\link{writeRSMParameters}()},
\code{\link{writeSelectionCHK}()}
}
\concept{RS Minerve IO}
