\name{riversCentralAsia-package}
\alias{riversCentralAsia-package}
\alias{riversCentralAsia}
\docType{package}
\title{
\packageTitle{riversCentralAsia}
}
\description{
\packageDescription{riversCentralAsia}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{riversCentralAsia}
\packageIndices{riversCentralAsia}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{riversCentralAsia}

Maintainer: \packageMaintainer{riversCentralAsia}
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
