% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacierBalance.R
\name{stepWiseGlacierBalance}
\alias{stepWiseGlacierBalance}
\title{Glacier volume change over time for small glaciers < 1km2}
\usage{
stepWiseGlacierBalance(M_mma, A_km2)
}
\arguments{
\item{M_mma}{matrix with glacier melt rates in mm/a (in columns) per
hydrological year (in rows). Also must contain a column named Hyear.}

\item{A_km2}{matrix with glacier area in km2 (in columns) for the
start year (the year 2000). Extracted from the geometries of the glaciers
in the RGI v6.0 data set.}
}
\value{
c(Q_m3a, V_km3, A_km2, Qimb_m3a) List of updated variables over the
  same time interval as M_mma. Q_m3a: glacier discharge, V_km3: glacier
  volume in water equivalents, A_km2: glacier area, Qimb_m3a: imbalance
  ablation (<0: glacier melt, >0: glacier increase)
}
\description{
Glacier volume change over time for small glaciers < 1km2
}
