% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacierBalanceP.R
\name{stepWiseGlacierBalancePerElBandP}
\alias{stepWiseGlacierBalancePerElBandP}
\title{Glacier volume change over time}
\usage{
stepWiseGlacierBalancePerElBandP(M_mma, P_mma, A_km2, V_km3)
}
\arguments{
\item{M_mma}{matrix with glacier melt rates in mm/a (in columns) per
hydrological year (in rows). Also must contain a column named Hyear.}

\item{P_mma}{matrix with precipitation rates in mm/a (in columns) per
hydrological year (in rows). Also must contain a column named Hyear.}

\item{A_km2}{matrix with area of elevation bands in km2 (in columns) for
the start year (the year 2000). Extracted from the geometries of the
glaciers in the RGI v6.0 data set.}

\item{V_km3}{matrix with volume of elevation bands in km3 (in columns) for
the start year (year 2000). Area of elevation bands A_km2 times the
mean thickness in the elevation bands extracted from Farinotti et al.}
}
\value{
c(Q_m3a, V_km3, A_km2) List of updated variables over the same time
  interval as M_mma
}
\description{
Glacier volume change over time
}
