% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-vignette_glacier_melt.R
\docType{data}
\name{vignette_glacier_melt}
\alias{vignette_glacier_melt}
\title{Example data for vignettes on glaciers}
\format{
A list of objects
\describe{
  \item{rgi}{sf: A subset of the Randolph Glacier Inventory 6.0 in the Atbashy river catchment.}
  \item{rgi_elbands}{sf: A subset of the Randolph Glacier Inventory 6.0 with glaciers larger than 1 km2 subdivided into elevation bands.}
  \item{basin}{sf: Boundary of the Atbashy basin}
  \item{dem}{raster: Digital elevation model for the Atbashy catchment}
  \item{glacier_thickness}{raster: Estimates of glacier thickness. }
  \item{glaciers_hugonnet}{tibble: Glacier thinning rates.}
  \item{hist_sim}{tibble: Temperature time series from CHELSA v2.1 for one glacier}
  \item{fut_sim}{list of tibbles: Temperature time series from CMIP6 for one glacier}
}
}
\source{
RGI6 data set, Karger et al., 2020, Farinotti et al., 2019, Hugonnet et al., 2021, Copernicus CMIP6, SRTMGL1v003, 2013.
}
\usage{
vignette_glacier_melt
}
\description{
Example data for vignettes on glaciers
}
\examples{
data(vignette_glacier_melt)
rgi <- vignette_glacier_melt$rgi
}
\keyword{datasets}
