% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgen_daily_PT.R
\name{wgen_daily_PT}
\alias{wgen_daily_PT}
\title{Compute stochastic daily weather time series using RMAWGEN and a climate scenario}
\usage{
wgen_daily_PT(param, station_data, station_subset, clim_scen, returnVAR)
}
\arguments{
\item{param}{List with RMAWGEN parameters (see ?RMAWGEN::RMAWGEN for mroe infomation)}

\item{station_data}{List with key station statistics, including STATION_NAMES, ELEVATION, STATION_LATLONG, LOCATION, PRECIPITATION, TEMPERATURE_MIN and TEMPERATURE_MAX. This is output of the function prepare_RMAWGEN_input_data()}

\item{station_subset}{Optional subset of station names that should be modeled. If NULL, the entire set as given by station_data$STATION_NAMES will be used.}

\item{returnVAR}{Return the var models? If this option is set to FALSE, the P_gen and T_gen files become more manageable.}

\item{climScen}{Optional climate scenario that conditions the generation of future stochastic weather variables on norm climate (output from GMCs) for the individual stations. If NULL, the simulated period will cover the observation periuod.}
}
\value{
Stochastic multi-site weather generator RMAWGEN precipitation and temperature models.
}
\description{
Wraper function to compute daily weather variables (precipitation, minimum and maximum temperatures) based on a climate scenario
}
\seealso{
Other Helper functions: 
\code{\link{aggregate_to_monthly}()},
\code{\link{assess_fc_qual}()},
\code{\link{convert2HYY}()},
\code{\link{cutRaster2Basin}()},
\code{\link{decadeMaker}()},
\code{\link{generateHourlyFromDaily_PT}()},
\code{\link{generateSeqDates}()},
\code{\link{hyear}()},
\code{\link{monDateSeq}()},
\code{\link{monthlyNorms_ClimateScenario}()},
\code{\link{posixct2rsminerveChar}()},
\code{\link{wgen_daily_summaryStats}()}
}
\concept{Helper functions}
