% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wgen_daily_summaryStats.R
\name{wgen_daily_summaryStats}
\alias{wgen_daily_summaryStats}
\title{wgen_daily_summaryStats}
\usage{
wgen_daily_summaryStats(PT_sim, param, station_data, station_subset, clim_scen)
}
\arguments{
\item{PT_sim}{PT_sim list (output of wgen_daily_PT())}

\item{param}{wgen_daily_PT (RMAWGEN) parameters}

\item{station_data}{Station ERA5 data prepared with riversCentralAsia::prepare_RMAWGEN_input_data() function.}

\item{station_subset}{Character vector with station names to provided summary statistics for. If is.null(station_subset)==TRUE, station_subset = param$station.}

\item{clim_scen}{Climate scenario under consideration which was also used in the generation of the simulated climate in PT_sim. If is.null(clim_scen)==TRUE, the observed (ERA5) values are compared with the simulated values.}
}
\value{
Lists with precipitation and temperature summary statistics for each station in the station_subset variable.
}
\description{
Function to generate precipitation and min./max. summary statistics tables from 1 stochastic daily climate realization. param$nscenario > 1 not yet supported.
}
\seealso{
Other Helper functions: 
\code{\link{aggregate_to_monthly}()},
\code{\link{assess_fc_qual}()},
\code{\link{convert2HYY}()},
\code{\link{cutRaster2Basin}()},
\code{\link{decadeMaker}()},
\code{\link{generateHourlyFromDaily_PT}()},
\code{\link{generateSeqDates}()},
\code{\link{hyear}()},
\code{\link{monDateSeq}()},
\code{\link{monthlyNorms_ClimateScenario}()},
\code{\link{posixct2rsminerveChar}()},
\code{\link{wgen_daily_PT}()}
}
\concept{Helper functions}
