% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/writeRSMParameters.R
\name{writeRSMParameters}
\alias{writeRSMParameters}
\title{Write parameters to RS Minerve parameter file}
\usage{
writeRSMParameters(parameters, outfilepath)
}
\arguments{
\item{parameters}{Tibble with parameters for each RS Minerve object. Multiple
parameter sets may be possible. See Details & Examples for more info on the
column requirements.}

\item{outfilepath}{Character string with path to file to be written.}
}
\value{
NULL for success.
}
\description{
Write parameters to RS Minerve parameter file
}
\details{
`parameters` contains 5 columns: \cr
  `Object` (character) identifying the RSMinerve object type (e.g. Station).
    See Vignette Parameters for more information on the objects available in
    RSMinerveR. \cr
  `Name` is the user specified name of the object in the RSMinerve model
    (e.g. Meteo station A). \cr
  `Parameters` names all parameters for each object (e.g. Zone, X \[m\],
    Y \[m\], etc.). See Vignette Parameters for more information on the
    parameters of the available objects in RSMinerveR. \cr
  `Values` contains the parameter values (e.g. "A", 4500, 3000). \cr
  `Parameter set` (optional) is used to differentiate between multiple
    parameter sets. The function will write
    length(unique(parameters$`Parameter set`)) parameter files, appending the
    value of Parameter set to the base file name. \cr
  The outfilepath is edited in the case where several parameter sets are written
}
\examples{
\dontrun{
filepath <- normalizePath(file.path("Tutorial_Parameters.txt"))
params <- readRSMParameters(filepath)
params <- params |>
mutate(Values = ifelse((Object == "GSM" & Parameters == "Kgl [1/d]"),
                        Values/2, Values))
writeRSMParameters(params, filepath)
}
}
\seealso{
Other RS Minerve IO: 
\code{\link{getChunkSize}()},
\code{\link{load_minerve_input_csv}()},
\code{\link{readDBCSV}()},
\code{\link{readForcingCSV}()},
\code{\link{readForcingSTR}()},
\code{\link{readRSMParameters}()},
\code{\link{readResultCSV}()},
\code{\link{readResultDST}()},
\code{\link{readSelectionCHK}()},
\code{\link{writeSelectionCHK}()}
}
\concept{RS Minerve IO}
