% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ChirchikRiverBasin}
\alias{ChirchikRiverBasin}
\title{Discharge, precipitation and temperature in the Chirchik river basin,
Uzbekistan.}
\format{
A data frame with 29892 rows and 14 variables:
\describe{
  \item{date}{Date: Date in decadal or monthly time steps.}
  \item{data}{num: Value of variable.}
  \item{norm}{num: Multi-year average of values.}
  \item{units}{chr: Unit of values. Can be m^3/s (m3s) for discharge, mm (mm)
               for precipitation or ºC (degC) for temperature.}
  \item{type}{Factor: Parameter measured. Can be "Q" for discharge, "P" for
              precipitation or "T" for temperature.}
  \item{code}{chr: Unique station identifier issued by Uzbek HydroMet.}
  \item{station}{chr: Station name.}
  \item{river}{chr: Name of river the station is related to.}
  \item{basin}{chr: The name of the basin the station is part of.}
  \item{resolution}{chr: Temporal resolution of data. Can be dec for data in
                    10-day intervals (decadal) or mon for monthly intervals.}
  \item{lon_UTM42}{num: Longitude of station in UTM42.}
  \item{lat_UTM42}{num: Latitude of station in UTM42.}
  \item{altitude_masl}{num: Altitude of station in meter above mean sea level.}
  \item{basinSize_sqkm}{num: Size of the basin in square kilometers (km2).}
}
}
\source{
Uzbek HydroMet 2021, Andrey Yakovlev
}
\usage{
ChirchikRiverBasin
}
\description{
A dataset containing the 10-day (decadal) or monthly data for discharge,
precipitation and temperature in the Chirchiq river basin, Uzbekistan.
}
\keyword{datasets}
