% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decadeMaker.R
\name{decadeMaker}
\alias{decadeMaker}
\title{Creates a vector of decadal (10 days) dates}
\usage{
decadeMaker(s, e, type)
}
\arguments{
\item{s}{starting date in YYYY-mm-dd format}

\item{e}{end date in YYYY-mm-dd format}

\item{type}{'start' creates starting decade dates, 'end' creates ending decade dates.}
}
\value{
A sequence of decadal dates
}
\description{
This function creates a decadal 10-days dates vector that allows date tagging
for decadal (10-days) time series. This type of timeseries is usually used for
hydro-meteorological data in the former Soviet Republics. The intra-months decades
can be configured as dates at the beginning or end of the decade.
}
\seealso{
Other Helper functions: 
\code{\link{aggregate_to_monthly}()},
\code{\link{assess_fc_qual}()},
\code{\link{convert2HYY}()},
\code{\link{cutRaster2Basin}()},
\code{\link{generateHourlyFromDaily_PT}()},
\code{\link{generateSeqDates}()},
\code{\link{hyear}()},
\code{\link{monDateSeq}()},
\code{\link{monthlyNorms_ClimateScenario}()},
\code{\link{posixct2rsminerveChar}()},
\code{\link{wgen_daily_PT}()},
\code{\link{wgen_daily_summaryStats}()}
}
\concept{Helper functions}
