% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downscale_ClimPred_monthly_BasinElBands.R
\name{downscale_ClimPred_monthly_BasinElBands}
\alias{downscale_ClimPred_monthly_BasinElBands}
\title{Load Tabular Hydro-Meteorological Data}
\usage{
downscale_ClimPred_monthly_BasinElBands(
  pathN,
  fileListSearchpattern,
  startY,
  endY,
  basinElBandsShape
)
}
\arguments{
\item{pathN}{Path to climate projection files.}

\item{fileListSearchpattern}{Search pattern in climate projection files to filter climate projection scenarios. E.g. "2006-2100" for the joined CHELSA climate projection files that include the 'ACCESS1-3','CMCC-CM' and 'MIROC5' models.}

\item{startY}{Starting year of the climate projections.}

\item{endY}{Last year of climate projections.}

\item{basinElBandsShape}{Basin shapefile with elevation bands and subbasins that are implemented in the the semi-distributed rainfall-runoff model.}
}
\value{
List with 4 elements, including downscaled monthly precipitation (pr), monthly mean minimum temperature (tasmean), monthly mean temperature (tasmean) and monthly mean maximum temperature (tasmax).
}
\description{
Loads csv files with tabular hydrometeorological data where years are in rows
and decades or months are in columns. The function automatically detects if
monthly or decadal (10-day) data is provided. The function automatically computes
the long-term norm of the data provided and returns a time aware tibble with
date, data, data norm columns and code columns.
}
