% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacierArea_RGIF.R
\name{glacierArea_RGIF}
\alias{glacierArea_RGIF}
\title{Calculates glacier area based on glacier volume}
\usage{
glacierArea_RGIF(volume_km3)
}
\arguments{
\item{volume_km3}{glacier volume in km3}
}
\value{
glacier area in km2
}
\description{
Scaling relationships derived from the RGI v6.0 region 13 data set, combined
with the glacier thickness derived by Farinotti et al. 2019.
}
\seealso{
\code{\link{glacierVolume_RGIF}}, \code{\link{glacierArea_Erasov}}

Other glacier functions: 
\code{\link{glacierArea_Erasov}()},
\code{\link{glacierImbalAbl}()},
\code{\link{glacierMelt_TI}()},
\code{\link{glacierVolume_Erasov}()},
\code{\link{glacierVolume_RGIF}()}
}
\concept{glacier functions}
