% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glacierVolume_RGIF.R
\name{glacierVolume_RGIF}
\alias{glacierVolume_RGIF}
\title{Calculates glacier volume based on glacier area}
\usage{
glacierVolume_RGIF(area_km2)
}
\arguments{
\item{area_km2}{glacier area in km2}
}
\value{
glacier volume in km3
}
\description{
Scaling relationships derived from the RGI v6.0 region 13 data set, combined
with the glacier thickness derived by Farinotti et al. 2019.
}
\details{
Assuming a normal distribution of the residuals, the relative
  uncertainty of the volume estimate is given as 2 times the standard
  deviation of the relative residuals which is equal to 31%.
}
\note{
The derivation of the empirical relationship as well as a comparison
  to the Erasov function is given in the book [Modelling of Hydrological Systems in Semi-Arid Central Asia](https://hydrosolutions.github.io/caham_book/glacier_modeling.html).
}
\seealso{
\code{\link{glacierArea_RGIF}}, \code{\link{glacierVolume_Erasov}}

Other glacier functions: 
\code{\link{glacierArea_Erasov}()},
\code{\link{glacierArea_RGIF}()},
\code{\link{glacierBalance}()},
\code{\link{glacierDischarge_HM}()},
\code{\link{glacierImbalAbl}()},
\code{\link{glacierMelt_TI}()},
\code{\link{glacierRMSE}()},
\code{\link{glacierTotalAblation_HM}()},
\code{\link{glacierVolume_Erasov}()}
}
\concept{glacier functions}
