% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyear.R
\name{hyear}
\alias{hyear}
\title{Calculates the hydrological year for a given date}
\usage{
hyear(date)
}
\arguments{
\item{date}{as_date or as.POSIXct}
}
\value{
Hyear as numeric
}
\description{
In Central Asia the hydrological year starts in October of the previous year.
}
\examples{
Hyear <- hyear(lubridate::as_date(c("2015-09-30", "2015-10-01", "2016-01-27")))
}
