% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadTabularData.R
\name{loadTabularData}
\alias{loadTabularData}
\title{Load Tabular Hydro-Meteorological Data}
\usage{
loadTabularData(
  fPath,
  fName,
  code,
  stationName,
  rName,
  rBasin,
  dataType,
  units
)
}
\arguments{
\item{fPath}{Path to the input file}

\item{fName}{File name (.csv file as input required)}

\item{code}{Hydrometeorological station code}

\item{stationName}{Hydrometeorological station name/location}

\item{rName}{Name of river}

\item{rBasin}{Name of basin}

\item{dataType}{Type of data, either `Q` (discharge data), `T` (temperature data) or `P` (precipitation data)}

\item{units}{Data units}
}
\value{
Time-aware tibble with relevant data
}
\description{
Loads csv files with tabular hydrometeorological data where years are in rows
and decades or months are in columns. The function automatically detects if
monthly or decadal (10-day) data is provided. The function automatically computes
the long-term norm of the data provided and returns a time aware tibble with
date, data, data norm columns and code columns.
}
