% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_RMAWGEN_input_data.R
\name{prepare_RMAWGEN_input_data}
\alias{prepare_RMAWGEN_input_data}
\title{Prepare RMAWGEN input files from RS MINERVE data file that contains hourly observations}
\usage{
prepare_RMAWGEN_input_data(rsMinerve_data_csv_tibble)
}
\arguments{
\item{rsMinerve_data_csv_tibble}{RS MINERVE loaded csv file - read.table(paste0(filePath,fileName),sep=',') -}
}
\value{
List with 4 elements, including downscaled monthly precipitation (pr), monthly mean minimum temperature (tasmean), monthly mean temperature (tasmean) and monthly mean maximum temperature (tasmax).
}
\description{
Based on the basin-specific RS MINERVE database csv file, the function computes input files for the RMAWGEN precipitation and temperature weather generator.
}
\seealso{
Other Pre-processing: 
\code{\link{biasCorrect_ERA5_CHELSA}()},
\code{\link{climateScenarioPreparation_RMAWGEN}()},
\code{\link{computeDiurnalTemperatureCycle}()},
\code{\link{doQuantileMapping_Tobi}()},
\code{\link{downscale_ClimPred_monthly_BasinElBands}()},
\code{\link{gen_HRU_Climate_CSV_RSMinerve}()},
\code{\link{gen_basinElevationBands}()},
\code{\link{generate_ERA5_Subbasin_CSV}()},
\code{\link{import_iEasyHydro_data}()},
\code{\link{loadTabularData}()},
\code{\link{plotNormDevHYY}()}
}
\concept{Pre-processing}
