\name{riversCentralAsia-package}
\alias{riversCentralAsia-package}
\alias{riversCentralAsia}
\docType{package}
\title{
\packageTitle{riversCentralAsia}
}
\description{
\packageDescription{riversCentralAsia}
}
\details{
The package includes functions for pre-processing of data for hydrological modelling (for example for the generation of hydrological response units for hydrological modelling or for bias correction of climate forcing), for post-processing of simulation results (for example the generation of flow duration curves which are used for determining the design flow of hydropower plants), for interfacing with RS Minerve, a free numerical hydrological-hydraulic simulation software by \link[http://www.crealp.ch]{CREALP} (for example reading and writing of model input and output files) and a rudimentary glacier balance model.
The uses of the functions are extensively demonstrated in the open-source course book \link[https://hydrosolutions.github.io/caham_book/]{Modeling of Hydrological Systems in Semi-Arid Central Asia} and in the vignette of the package.
}
\author{
\packageAuthor{riversCentralAsia}

Maintainer: \packageMaintainer{riversCentralAsia}
}
\references{
T. Siegfried and B. Marti (2022): Modeling of Hydrological Systems in Semi-Arid Central Asia. URL: \link[https://hydrosolutions.github.io/caham_book/]{https://hydrosolutions.github.io/caham_book/}, DOI: \link[https://doi.org/10.5281/zenodo.6349983]{https://doi.org/10.5281/zenodo.6349983}.
}
\keyword{ package }
